+++ src/upm/iPressureSensor.h
@@ -9,7 +9,7 @@ namespace upm
    {
    public:
        virtual mraa_result_t getPressure (int32_t *value) = 0;
-       virtual mraa_result_t getTemperature (float *value) = 0;
+       virtual mraa_result_t getTemperature (int32_t *value) = 0;
        virtual bool isConfigured() = 0;
        virtual bool isAvailable() = 0;
        virtual ~IPressureSensor() {}
